feat(front-kit-3): extracted orderlistLayout
