+++ packages/falcon-ui-kit/src/Order/OrderListLayout.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { themed, Box } from '@deity/falcon-ui';
+
+export const OrderListLayout = themed({
+  tag: Box,
+  defaultTheme: {
+    orderListLayout: {
+      display: 'flex',
+      flexDirection: 'column'
+    }
+  }
+});
