+++ packages/falcon-ui-kit/src/Order/index.ts
@@ -2,3 +2,4 @@ export * from './OrderLayout';
 export * from './OrderItemSummary';
 export * from './OrderItemSummaryLayout';
 export * from './EmptyOrderList';
+export * from './OrderListLayout';
